+++ Engine/Source/Base.h
@@ -24,4 +24,4 @@ typedef unsigned long long CeU64;
 #endif
 
 /// @brief MAJOR.MINOR.PATCH - DD.MM.YYYY
-#pragma CE_VERSION "0.0.1 - 27.02.2024"
+#pragma CE_VERSION "0.0.2 - 27.02.2024"