+++ Engine/Source/Platform/FileSystem.h
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "Base.h"
+
+#include <string>
+#include <fstream>
+
+namespace Core
+{
+    class CE_API FileSystem
+    {
+    public:
+        FileSystem(){};
+        ~FileSystem(){};
+
+        static std::string ReadFileContent(const std::string &filename);
+    };
+
+}